migration_server/restfulapi/requirements.txt
Flask
paho-mqtt==1.6.1
Pillow
pandas
influxdb